Completion of major change but with a single problem
New method of map creation is done and responsiveness, text angle, load animation and data changes are done. But the transform property is faulty in ios device (except mac). Its rendering the html tags of svg that are inside foreignObject at (0,0) coordinates of the svg and not at the center as expected.